add back in algorithm arguments as defaults as well
